Simplify code of http request
