--- lazyllm/tools/http_request/http_executor.py
@@ -1,80 +1,0 @@
-import httpx
-
-class HttpExecutorResponse:
-    headers: dict[str, str]
-    response: httpx.Response
-
-    def __init__(self, response: httpx.Response = None):
-        self.response = response
-        self.headers = dict(response.headers) if isinstance(self.response, httpx.Response) else {}
-
-    @property
-    def is_file(self) -> bool:
-        """
-        check if response is file
-        """
-        content_type = self.get_content_type()
-        file_content_types = ['image', 'audio', 'video']
-
-        return any(v in content_type for v in file_content_types)
-
-    def get_content_type(self) -> str:
-        return self.headers.get('content-type', '')
-
-    def extract_file(self) -> tuple[str, bytes]:
-        """
-        extract file from response if content type is file related
-        """
-        if self.is_file:
-            return self.get_content_type(), self.body
-
-        return '', b''
-
-    @property
-    def content(self) -> str:
-        if isinstance(self.response, httpx.Response):
-            return self.response.text
-        else:
-            raise ValueError(f'Invalid response type {type(self.response)}')
-
-    @property
-    def body(self) -> bytes:
-        if isinstance(self.response, httpx.Response):
-            return self.response.content
-        else:
-            raise ValueError(f'Invalid response type {type(self.response)}')
-
-    @property
-    def status_code(self) -> int:
-        if isinstance(self.response, httpx.Response):
-            return self.response.status_code
-        else:
-            raise ValueError(f'Invalid response type {type(self.response)}')
-
-    @property
-    def size(self) -> int:
-        return len(self.body)
-
-    @property
-    def readable_size(self) -> str:
-        if self.size < 1024:
-            return f'{self.size} bytes'
-        elif self.size < 1024 * 1024:
-            return f'{(self.size / 1024):.2f} KB'
-        else:
-            return f'{(self.size / 1024 / 1024):.2f} MB'
-
-class HttpExecutor:
-    def __init__(self, method, url, headers, params, body):
-        self.method = method
-        self.url = url
-        self.headers = headers
-        self.params = params
-        self.body = body
-
-    def invoke(self):
-        if self.method in ('get', 'head', 'post', 'put', 'delete', 'patch'):
-            response = httpx.request(method=self.method, url=self.url, headers=self.headers, params=self.params,
-                                     data=self.body)
-            return HttpExecutorResponse(response)
-        raise ValueError(f'Invalid http method {self.method}')
