+++ tests/basic_tests/test_http_node.py
@@ -1,10 +1,9 @@
-from lazyllm.tools.http_request.http_request import HttpRequest, NodeExecutionStatus
+from lazyllm.tools.http_request.http_request import HttpRequest
 
 class TestHTTPRequest(object):
 
     def test_http_request(self):
         http_request = HttpRequest('get', 'https://httpbin.org/ip', api_key='', headers={}, params={}, body='')
         r = http_request()
-        assert r['status'] == NodeExecutionStatus.SUCCEEDED
-        assert r['output']['status_code'] == 200
-        assert 'origin' in r['output']['content']
+        assert r['status_code'] == 200
+        assert 'origin' in r['content']
